requirements.txt
pyyaml>=6.0
dotenv>=0.9.9